New approach to git aliases
For the last few years, I've been forcing myself to use vim-fugitive,
even though the learning curve seemed quite steep. I started using it
for a very narrow set of actions and then slowly, slowly expanding over
time. By now, I have internalized a significant portion of fugitive's
functionality, but between its mappings and all of my git aliases, it's
starting to feel like I need to hold too many things in my head.

This change is an attempt to alleviate the problem. The approach I took
is to configure the shell to use the same mappings that fugitive uses.
*Exactly* means that not every abbreviation starts with the keystroke
`g`. Radical, I know. Maybe I'll hate it tomorrow, but for now, I'm
willing to give it a try.

And yes, abbreviations, not aliases. The neat thing about fish
abbreviations is that they expand to the full command, which is
excellent for at least two reasons that I can think of:

1) It is always self-evident which command you're running. It makes it
super easy for people who are pairing or screen-sharing with you to see
what you're doing.
2) Makes searching through your shell's history super-efficient. Since
it makes sure you use the same command every time (as opposed to
switching between the alias and the full command, like if you copy/paste
some code from the internet, for example), all the history entries will
be consistent.